[#1244] Factor out command execution logic into common cli library
